% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{inner_product_phi}
\alias{inner_product_phi}
\title{Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----}
\usage{
inner_product_phi(X, v, i, j)
}
\arguments{
\item{X, }{matrix n by p}

\item{v, }{contrast vector n by 1}

\item{i, }{first index}

\item{j, }{second index}
}
\value{
parameters: a, b, c the coefficients of the quadratic equation such that (ax^2 + bx + c <= 0)
}
\description{
Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----
}
\keyword{internal}
