% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{kmeans_compute_S_genCov}
\alias{kmeans_compute_S_genCov}
\title{Compute set S for isotropic case}
\usage{
kmeans_compute_S_genCov(
  X,
  estimated_k_means,
  all_T_clusters,
  all_T_centroids,
  n,
  XTv,
  XTv_norm,
  dir_XTv,
  v_vec,
  v_norm,
  T_length,
  Sig_XTv_norm,
  k
)
}
\description{
Compute set S for isotropic case
}
