% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{norm_sq_phi}
\alias{norm_sq_phi}
\title{Represent ||x'(phi)_i-x'(phi)_j||_2^2 as a quadratic function in phi ----}
\usage{
norm_sq_phi(X, v, XTv, XTv_norm, dir_XTv, v_norm, i, j)
}
\arguments{
\item{v, }{the vector}

\item{XTv, }{vector p by 1}

\item{XTv_norm, }{norm of XTv}

\item{dir_XTv, }{vector p by 1 := XTv/XTv_norm}

\item{v_norm, }{2-norm of vector v}

\item{i, }{first index}

\item{j, }{second index}
}
\value{
parameters: a, b, c the coefficients of the quadratic equation
such that (ax^2 + bx + c <= 0)
}
\description{
Represent ||x'(phi)_i-x'(phi)_j||_2^2 as a quadratic function in phi ----
}
\keyword{internal}
