% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_union.R
\name{summary.kmeans_inference_union}
\alias{summary.kmeans_inference_union}
\title{Summarize the union inferential result for k-means clustering}
\usage{
\method{summary}{kmeans_inference_union}(object, ...)
}
\arguments{
\item{object}{output from running \code{kmeans_inference_union}}

\item{...}{to be passed to methods}
}
\value{
A data frame with summarized results.
}
\description{
Summarize the union inferential result for k-means clustering.
}
