% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_inf.R
\name{kmeans_estimation}
\alias{kmeans_estimation}
\title{k means estimation}
\usage{
kmeans_estimation(
  X,
  k,
  iter.max = 10,
  nstart = 1,
  seed = 1234,
  tol_eps = 1e-06,
  verbose = TRUE
)
}
\description{
k means estimation
}
