% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_inf.R
\name{kmeans_inference}
\alias{kmeans_inference}
\title{k means inference}
\usage{
kmeans_inference(
  X,
  k,
  cluster_1,
  cluster_2,
  sig,
  iter.max = 10,
  nstart = 1,
  seed = 1234,
  tol_eps = 1e-06,
  verbose = TRUE
)
}
\description{
k means inference
}
