% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{norm_phi_canonical_kmeans}
\alias{norm_phi_canonical_kmeans}
\title{Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----}
\usage{
norm_phi_canonical_kmeans(
  X,
  last_centroids,
  XTv,
  XTv_norm,
  dir_XTv,
  v_norm,
  cl,
  k,
  v,
  i
)
}
\arguments{
\item{XTv, }{vector p by 1}

\item{XTv_norm, }{norm of XTv}

\item{dir_XTv, }{vector p by 1 := XTv/XTv_norm}

\item{v_norm, }{2-norm of vector v}

\item{cl, }{factor vector n by 1 (most recent cluster assignment)}

\item{k, }{cluster of interest}

\item{v, }{contrast vector n by 1}

\item{i, }{index of observation}
}
\value{
parameters: a, b, c the coefficients of the quadratic equation such that (ax^2 + bx + c <= 0)
}
\description{
Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----
}
\keyword{internal}
